Tests marrays directly
Thanks to Malcolm Reynolds and Tom Hermann, lisp-unit:numerical-equal
is now a generic function, so in the new file test-unit/augment.lisp,
a method has been defined to work on marrays.  This now permits the
direct comparison in tests of marrays, rather than looking at their CL
equivalent with cl-array.  So far the only test changed is
vector-add.