Complex functions
Additional definition of functions of complex numbers and/or returning
complex numbers.  Most of these duplicate functions that Lisp has
already, so everything is defined with a "cx-" prefix.  No effort has
been made to unify the interface; this requires some thought.  A fix
has been made to body-expand; previously, it would not properly
compile a function that returned a structure by value but didn't have
structures by value as arguments.