Simulated annealing functioning
Simulated annealing now works.  The trivial example given in GSL
source doc/examples/siman.c is coded up at the end of
simulated-annealing.lisp.  Both versions produce the same
answer,
     1.36313001299
#<VECTOR-DOUBLE-FLOAT #(1.3631300129927695d0)>
Note that the parameters for this example are slightly different than
what is given in the manual; the manual version of the parameters are
commented out in #'simulated-annealing-example.
The Lisp version takes about 26.5 seconds, the C version about 0.85
seconds.