Simulated annealing loaded in defsystem
Simulated annealing is now loaded with the defsystem, and its presence
advertised in documentation/index.html.  The file
simulated-annealing.lisp has been moved to solve-minimize-fit.
The GSL "make check" tests have been added at the end of the file, but
they are not added to GSLL's examples/tests yet because of how slow it
runs.