Use CFFI-grovel to define enumerations and programming constants
Instead of hardwiring GSL's #define or enum values, define them in
init/libgsl-unix.lisp and run CFFI-grovel on them.  This has the
advantage that if these values change in GSL, GSLL will automatically
get the new values; also, there was an error in the 'integrate-method
enum due to an incorrect starting value.  It has the disadvantage that
building GSLL now requires a C compiler and some of the header (.h)
files for GSL.  Also, it's not clear what the form for libgsl-*.lisp
files are for non-unix OSes.