Grovel physical constants using new CFFI-grovel
Stelian Ionescu has modified CFFI-grovel to generate double-float
constants with :type double-float in constant.  GSLL has been ported
to use that.  This requires a version of CFFI which includes the patch
of Mon Jun 8 10:29:44 EDT 2009.  It works, but until this has a
release number, it will stay out of the master GSLL branch.