Hand expansion of project-ntuple; start writing ntuple example
In order to avoid or defer the complicated process of modifying
defmfun definitions to handle two cbstructs, the definition of
project-ntuple has been written by hand as a defun etc., by expanding
a simpler defmfun and making the necessary modifications and mergers
with an editor.  It has not yet been tested.  The test given in GSL
ntuple/test.c is partially ported to Lisp.