Constants defined for conditions with eval-when
The constants used for GSL conditions (+EDOM+ etc.) are defined inside
an eval-when so that implementations like CLISP that don't evaluate at
compile time will still be able to use the constants in the following
parts of the file.  This now compiles in CLISP (as well as SBCL and
CCL).