Multiple foreign functions in the reinitialize-instance for mobjects
It is now possible to specify two foreign functions for the
reinitialize-instance for defmobject.  The switch between the function
depends on the presence or absence of optional/key arguments.  This is
used in the definition of one-dimensional-minimizer to switch between
gsl_min_fminimizer_set and gsl_min_fminimizer_set_with_values
depending on the non-NIL value of f-minimum.  The set-with-values case
has not yet been tested.