Cholesky: add tests; return argument for matrix-product
Added eight tests from GSL's test for Cholesky decomposition and
solving.  The four decomposition tests incorrectly produce zero in
CCL.  The first optional argument in matrix-product and
matrix-product-symmetric can now be nil, to indicate a matrix of the
right dimensions should be constructed.  This is the default; the
default previously was to make that matrix directly, which meant if
any other optional arguments were supplied, the complicated default
construction for the return matrix would have to be re-entered.