Added symbol-macrolets to most of the do-matrix-* macros.
This means the symbol 'elm' can be used within the body given
to these macros and it will expand into a (maref matrix i j) and
so can be used to get or setf that element. I am yet to figure out
how to add a real macro to do-matrix-[cols,rows] which will take
one argument, so that

(do-matrix-rows (mtx r)
  (princ (elm 1)))

will print the second element of each row. There has to be a way...